审核子内容bug修复

+++ publiccms-parent/publiccms-core/src/main/java/com/publiccms/controller/admin/cms/CmsContentAdminController.java
@@ -279,7 +279,7 @@ public class CmsContentAdminController {
                     categoryIdSet.add(entity.getCategoryId());
                 }
             }
-            for (CmsContent parent : service.getEntitys(parentIdSet.toArray(new Integer[parentIdSet.size()]))) {
+            for (CmsContent parent : service.getEntitys(parentIdSet.toArray(new Long[parentIdSet.size()]))) {
                 publish(site, parent, admin);
             }
             for (CmsCategory category : categoryService.getEntitys(categoryIdSet.toArray(new Integer[categoryIdSet.size()]))) {
